[Silver V] Title: 최소공배수, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/13241. 최소공배수/최소공배수.cc
@@ -3,7 +3,7 @@ using namespace std;
 
 typedef long long int lli;
 
-int gcd(long long int a, lli b) {
+int gcd(lli a, lli b) {
 	if (b == 0) return a;
 	return gcd(b, a % b);
 }
